backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==4.2.4
django-cors-headers==4.3.0
django-extensions==3.2.3
django-filter==23.2
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.5
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
msgpack==1.0.8
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
psycopg2-binary==2.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
rpds-py==0.12.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2023.3
uritemplate==4.1.1
zope.interface==6.2
